requirements.txt
fastapi==0.115.8
Flask==3.1.0
flask-cors==5.0.1
uvicorn==0.34.0
requests
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
torch==2.5.1
torchvision==0.20.1
ultralytics==8.3.96
pandas==2.2.3
scipy==1.15.2
supervision==0.25.1
matplotlib==3.10.1
seaborn==0.13.2
SQLAlchemy==2.0.38
psutil==7.0.0
bitcoinlib==0.7.2
python-bitcoinlib==0.12.2
pycryptodome==3.21.0
pydantic==2.10.6
python-multipart==0.0.20
tqdm==4.67.1
